+++ lib/td_auth/canada/abilities.ex
@@ -1,15 +1,11 @@
 defmodule TdBg.Canada.Abilities do
   @moduledoc false
-  alias TdAuth.Accounts.Group
-  alias TdAuth.Accounts.User
   alias TdAuth.Auth.Claims
   alias TdAuth.Permissions.AclEntry
   alias TdAuth.Permissions.Role
   alias TdCache.Permissions
 
   defimpl Canada.Can, for: Claims do
-    @notification_groups ["business_glossary_view", "business_glossary_management"]
-
     # administrator is superpowerful
     def can?(%Claims{is_admin: true}, _action, _resource), do: true
 
@@ -30,16 +26,6 @@ defmodule TdBg.Canada.Abilities do
 
     def can?(claims, :delete, %{resource_type: "domain", resource_id: domain_id}) do
       authorized?(claims, :delete_acl_entry, domain_id)
-    end
-
-    def can?(%Claims{jti: jti, groups: groups}, :view, Group) do
-      Permissions.has_any_permission_on_resource_type?(jti, [:create_acl_entry], "domain") or
-        Enum.any?(@notification_groups, &(&1 in groups))
-    end
-
-    def can?(%Claims{jti: jti, groups: groups}, :view, User) do
-      Permissions.has_any_permission_on_resource_type?(jti, [:create_acl_entry], "domain") or
-        Enum.any?(@notification_groups, &(&1 in groups))
     end
 
     def can?(%Claims{jti: jti}, :view, Role) do